# public/api/v1/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SGFEP RESTful API",
    "description": "البوابة الموحدة لتكامل الأنظمة والربط الرقمي لمنصة SGFEP (وزارة التكوين والتعليم المهنيين). تدعم المصادقة المزدوجة عبر مفاتيح API وعلامات Bearer JWT.",
    "version": "1.0.0",
    "contact": {
      "name": "إدارة المعلوماتية - وزارة التكوين والتعليم المهنيين",
      "email": "support@mfep.gov.dz"
    }
  },
  "servers": [
    {
      "url": "/sig",
      "description": "خادم XAMPP المحلي أو المسار الفرعي"
    },
    {
      "url": "/",
      "description": "خادم التطوير المباشر أو خادم الإنتاج الرئيسي"
    }
  ],
  "paths": {
    "/api/v1/auth/token": {
      "post": {
        "summary": "طلب رمز JWT جديد",
        "description": "تبادل مفتاح API Key (سواء كان sgfep_live_... للحسابات أو sgfep_ext_... للمنصات الخارجية) للحصول على رمز Bearer JWT نشط وصالح للوصول الآمن.",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "api_key": {
                    "type": "string",
                    "description": "مفتاح API الخاص بك للربط",
                    "example": "sgfep_ext_8c6976e5b5410415bde908bd4dee15df"
                  }
                },
                "required": [
                  "api_key"
                ]
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "api_key": {
                    "type": "string",
                    "description": "مفتاح API الخاص بك للربط",
                    "example": "sgfep_ext_8c6976e5b5410415bde908bd4dee15df"
                  }
                },
                "required": [
                  "api_key"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم إصدار رمز JWT بنجاح",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "code": 200,
                  "message": "JWT Token issued successfully.",
                  "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0L3NpZyIsImlhdCI6MTc4MTQyOTgwOCwiZXhwIjoxNzgxNDU4NjA4LCJzdWIiOiJjbGllbnRfMSIsInVzZXJuYW1lIjoiYXBpX21paG5hdGkiLCJyb2xlIjoiYXBpX2clientIn0.xxx",
                  "token_type": "Bearer",
                  "expires_in": 28800,
                  "user": {
                    "username": "api_mihnati",
                    "name": "Mihnati Integration"
                  }
                }
              }
            }
          },
          "400": {
            "description": "طلب غير صحيح (معلمة مفقودة)",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "code": 400,
                  "message": "Bad Request. Missing api_key parameter."
                }
              }
            }
          },
          "401": {
            "description": "مفتاح API غير صحيح أو موقوف",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "code": 401,
                  "message": "Invalid API key or inactive user."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "أرسل الرمز المستخرج من /api/v1/auth/token في رأس الطلب بصيغة: Authorization: Bearer <jwt_token>"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "أرسل مفتاح الـ API المباشر في رأس الطلب بصيغة: X-API-Key: sgfep_ext_..."
      }
    }
  }
}